1. Cleanup 2. Logging + error handling added to enc_info_plainfile_storage

+++ storage/rocksdb/enc_info_plainfile_storage.cc
@@ -1,12 +1,16 @@
 #include "./enc_info_plainfile_storage.h"
 #include "rocksdb/file_system.h"
 #include "rocksdb/util/crc32c.h"
+#include "rocksdb/logging/logging.h"
 
 namespace myrocks {
+using namespace rocksdb;
+
 static constexpr uint32_t g_serializtion_format_version = 1;
 static constexpr const char *g_serialization_format =
   "Version: %u\nMasterKeyId: %u\nMasterKeyRotationInProgress: %u\nServerUuid: %s\nCRC: %u\n";
-static constexpr size_t uint32_t_strlen = strlen("4294967295");
+static constexpr const char *g_CRC_item = "CRC: ";
+static constexpr size_t uint32_t_strlen = strlen("4294967295");  // uint32_t max
 static constexpr size_t uuid_strlen = strlen("9ddb9c4d-c4bf-4db3-9303-0e52470a484f");
 static constexpr size_t g_serialization_buffer_size = strlen(g_serialization_format)
                                                       + uint32_t_strlen // version
@@ -15,16 +19,18 @@ static constexpr size_t g_serialization_buffer_size = strlen(g_serialization_for
                                                       + uuid_strlen     // uuid
                                                       + uint32_t_strlen // CRC
                                                       + 1;              // '0'
-static char g_serializtion_buffer[g_serialization_buffer_size];
+static char g_serialization_buffer[g_serialization_buffer_size];
 
 EncryptionInfoPlainFileStorage::EncryptionInfoPlainFileStorage(
     const std::string& filePath,
     const std::shared_ptr<rocksdb::FileSystem> fs,
-    const std::string& uuidHint)
+    const std::string& uuidHint,
+    std::shared_ptr<rocksdb::Logger> logger)
 : filePath_(filePath),
   backupFilePath_(filePath_+".bak"),
   fs_(fs),
-  uuidHint_(uuidHint)
+  uuidHint_(uuidHint),
+  logger_(logger)
 {
     std::unique_lock<std::mutex> lock(fileAccessMtx_);
     recover();
@@ -38,7 +44,6 @@ void EncryptionInfoPlainFileStorage::StoreCurrentMasterKeyId(uint32_t Id)
     serialize(filePath_, encryptionInfo_);
 }
 
-    // returns 0 if no id stored
 uint32_t EncryptionInfoPlainFileStorage::GetCurrentMasterKeyId()
 {
     std::unique_lock<std::mutex> lock(fileAccessMtx_);
@@ -77,7 +82,7 @@ void EncryptionInfoPlainFileStorage::serialize(const std::string &filePath,
       rocksdb::Slice data;
       std::unique_ptr<rocksdb::FSWritableFile> file;
 
-      snprintf(g_serializtion_buffer, g_serialization_buffer_size, g_serialization_format,
+      snprintf(g_serialization_buffer, g_serialization_buffer_size, g_serialization_format,
         g_serializtion_format_version,
         info->CurrentMasterKeyId,
         info->MasterKeyRotationInProgress,
@@ -85,19 +90,39 @@ void EncryptionInfoPlainFileStorage::serialize(const std::string &filePath,
         0);
 
       // crc
-      char* ptr = strstr(g_serializtion_buffer, "CRC: ");
-      if (ptr) ptr += strlen("CRC: ");
-      uint32_t crc = rocksdb::crc32c::Extend(0, g_serializtion_buffer,
-                                         ptr - g_serializtion_buffer);
+      char* ptr = strstr(g_serialization_buffer, g_CRC_item);
+      if (ptr) ptr += strlen(g_CRC_item);
+      uint32_t crc = rocksdb::crc32c::Extend(0, g_serialization_buffer,
+                                         ptr - g_serialization_buffer);
 
       sprintf(ptr, "%u\n", crc);
-      data = rocksdb::Slice(g_serializtion_buffer);
+      data = rocksdb::Slice(g_serialization_buffer);
 
-      fs_->NewWritableFile(filePath, rocksdb::FileOptions(), &file, nullptr);
-      file->Append(data, rocksdb::IOOptions(), nullptr);
-      file->Close(rocksdb::IOOptions(), nullptr);
+      auto status = fs_->NewWritableFile(filePath, rocksdb::FileOptions(), &file, nullptr);
+      if (!status.ok()) {
+        ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::serialize(). Failed to create %s file.",
+            filePath.c_str());
+        return;
+      }
+      status = file->Append(data, rocksdb::IOOptions(), nullptr);
+      if (!status.ok()) {
+        ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::serialize(). Failed to append to %s file.",
+            filePath.c_str());
+        return;
+      }
+      status = file->Close(rocksdb::IOOptions(), nullptr);
+      if (!status.ok()) {
+        ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::serialize(). Failed to close %s file.",
+            filePath.c_str());
+        return;
+      }
       if (backup) {
-        fs_->DeleteFile(backupFilePath_, rocksdb::IOOptions(), nullptr);
+        status = fs_->DeleteFile(backupFilePath_, rocksdb::IOOptions(), nullptr);
+        if (!status.ok()) {
+          ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::serialize(). Failed to delete %s file.",
+            filePath.c_str());
+          return;
+        }
       }
     }
 }
@@ -111,7 +136,12 @@ EncryptionInfoPlainFileStorage::deserialize(const std::string &filePath)
 
     auto status = fs_->NewSequentialFile(filePath, rocksdb::FileOptions(), &file, nullptr);
     if (status.ok()){
-        file->Read(g_serialization_buffer_size, rocksdb::IOOptions(), &data, g_serializtion_buffer, nullptr);
+        status = file->Read(g_serialization_buffer_size, rocksdb::IOOptions(), &data, g_serialization_buffer, nullptr);
+        if (!status.ok()) {
+            ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::deserialize(). Reading from file %s failed.",
+              filePath.c_str());
+            return nullptr;
+        }
         char uuid[128];
         uint32_t version;
         sscanf(data.data(), g_serialization_format,
@@ -122,15 +152,24 @@ EncryptionInfoPlainFileStorage::deserialize(const std::string &filePath)
                 &newInfo->CRC);
 
         // crc check
-        char *ptr = strstr(g_serializtion_buffer, "CRC: ");
-        if (!ptr) return nullptr;
-        ptr += strlen("CRC: ");
+        char *ptr = strstr(g_serialization_buffer, g_CRC_item);
+        if (!ptr) {
+            ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::deserialize(). Failed to read CRC.");
+            return nullptr;
+        }
+        ptr += strlen(g_CRC_item);
         uint32_t crc;
         sscanf(ptr, "%u\n", &crc);
-        if (newInfo->CRC != crc) return nullptr;
+        if (newInfo->CRC != crc) {
+            ROCKS_LOG_WARN(logger_, "EncryptionInfoPlainFileStorage::deserialize(). CRC mismatch for file %s",
+             filePath.c_str());
+            return nullptr;
+        }
 
         newInfo->ServerUuid = std::string(uuid);
         if (g_serializtion_format_version != version) {
+            ROCKS_LOG_ERROR(logger_, "EncryptionInfoPlainFileStorage::deserialize() failed. Serialization format mismatch. Current: %u, expected: %u",
+              version, g_serializtion_format_version);
             return nullptr;
         }
     } else {
@@ -150,7 +189,10 @@ void EncryptionInfoPlainFileStorage::recover() {
         // If the main file is not OK, recover from backup.
         auto mainInfo = deserialize(filePath_);
         if (mainInfo) {
-            fs_->DeleteFile(backupFilePath_, rocksdb::IOOptions(), nullptr);
+            status = fs_->DeleteFile(backupFilePath_, rocksdb::IOOptions(), nullptr);
+            if (!status.ok()){
+                ROCKS_LOG_ERROR(logger_, "EncryptionInfoPlainFileStorage::recover(). Failed deleting backup file.");
+            }
             return;
         }
 
@@ -161,7 +203,10 @@ void EncryptionInfoPlainFileStorage::recover() {
             // as we are modifying main and backup files independently.
         }
         serialize(filePath_, backupInfo, false);
-        fs_->DeleteFile(backupFilePath_, rocksdb::IOOptions(), nullptr);
+        status = fs_->DeleteFile(backupFilePath_, rocksdb::IOOptions(), nullptr);
+        if (!status.ok()){
+            ROCKS_LOG_ERROR(logger_, "EncryptionInfoPlainFileStorage::recover(). Failed deleting backup file.");
+        }
     }
 }
 } // namespace myrocks
